Add CompositeEventSourceBuilder
We currently have MergedEventSource to compose multiple event sources
with the same event type, but it’s unergonomic because it takes an array
of event sources, but you can’t construct an array of heterogeneous
event sources due to limitations on protocols with associated types.

The current workaround is to explicitly wrap the members of the array in
AnyEventSource, but explicit use of type erasure in clients is
unpleasant.

Using a builder lets us take each individual event source as a generic
parameter and handle the type erasure internally.

I called it CompositeEventSourceBuilder rather than MergedEventSource by
analogy to CompositeDisposable.