Remove dispatch queues and internal locking from MobiusLoop
Mobius now dispatches events synchronously, and runs effects synchronously on
the same thread. Concurrent access is unsafe, and is guarded against in debug
builds.

This does mean it's possible to create infinite recursion in a Mobius loop.
However, the corresponding code in the old model would lead to infinite effort
and context switching, which is unlikely to be better.