Handle events and effects in order, without reentrancy hazards
The ordering guarantee we want to commit to at this time is: if two
events E1 and E2 are dispatched, all effects produced by E1 will be
handled before any effects produced by E2.

(Note that an effect is considered “handled” when the effect handler’s
accept returns. The effect handler may still be doing work
asynchronously.)

The easiest way to provide this guarantee is to fully sequence all
events and effects. Combined with the use of arrays for effects that
means execution order is fully deterministic and in the obvious
order, but we’re not committing to that right now.