Disable ConnectableClass assertion
We've disabled this in production because it's not critical and was being hit occasionally.
We have a regression test which _was_ hitting it 100% of the time, but is now hitting it
about 0.05% of the time. Since use of ConnectableClass for effect handlers is deprecated,
we don't want to spend the time to track down the cause.