Fix logical race caused by asynchronous view connection cleanup
View connections are disposed asynchronously on the view queue (typically
the main queue). If an event is dispatched on the view queue after
stopping the Mobius controller, we would hit an assertion. Unlike
previous logic races, this could reasonably be worked around in client
code, but it would introduce annoying coupling and synchronization work
which we can handle generically using our existing dropping-events-on-
the-floor technique.